Updates usage of stft and solve
The short-time Fourier transform used to be part of the torchaudio
package, but has now been moved into the main torch package. This
commit updates the code to use the new location.

Similarly, solve has been moved into linalg.solve, and has had its
arguments reversed and output changed. This commit applies these
changes.

In the future it will be necessary to use return_complex=True with
the STFT, and use complex numbers rather than tensors with two
components, but this change has been postponed.

Thanks to @YuanZi1501040205 for prompting this change in Issue #1.